--- mrtt-ui/src/components/Footer/index.js
@@ -1,2 +1,0 @@
-import Footer from './Footer'
-export default Footer
