+++ mrtt-ui/src/components/MobileFooter/MobileFooter.test.js
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react'
 import { composeStories } from '@storybook/testing-react'
 
-import * as stories from './Footer.stories'
+import * as stories from './MobileFooter.stories'
 
 const { Primary } = composeStories(stories)
