+++ mrtt-ui/src/components/MobileFooter/index.js
@@ -1,0 +1,2 @@
+import MobileFooter from './MobileFooter'
+export default MobileFooter
